[tools] perf hists: Do not link a pair if already linked
Message-id: <20200415143309.16001-156-mpetlan@redhat.com>
Patchwork-id: 302399
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 156/469] perf hists: Do not link a pair if already linked
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit 7d1a5efa20dbfea97cb93b99c67ce5cd5c4a4dbc
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Aug 7 10:45:30 2019 -0300

description
===========
When we have multiple events in a group we link hist_entries in the
non-leader evsel hists to the one in the leader that points to the same
sorting criteria, in hists__match().

For 'perf report' we do this just once and then print the results, but
for 'perf top' we need to look if this was already done in the previous
refresh of the screen, so check for that and don't try to link again.

This is part of having 'perf top' using the hists browser for showing
multiple events in multiple columns.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>