[tools] perf hist: Remove dummy entries when finding real ones
Message-id: <20200415143309.16001-158-mpetlan@redhat.com>
Patchwork-id: 302401
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 158/469] perf hist: Remove dummy entries when finding real ones.
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit 5f8b4d5d237a3e2e35509da4e63769ae5c82c085
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Fri Aug 9 17:56:06 2019 -0300

description
===========
When he have an event group we have multiple struct hist instances, one
per evsel, and in each of these hists we may have hist_entries that
point to the same thing being observed, say a symbol, i.e. if we're
looking at instructions and cycles, then we'll have one hist_entry in
the "instructions" evsel and another in the "cycles" evsel.

We need to link those to then show one column for each. When we're
looking at some other pair of events, say instructions and cache misses,
we may have just the "instructions" hist entry and not one for "cache
misses", as instructions not necessarily generate cache misses, as the
logic expects one hist_entry per evsel, we end up adding "dummy"
hist_entries.

This is enough for 'perf report', that does this matching operation
(hists__match()) just once after processing all events, but for 'perf
top', we do this at each refresh, so we may finally find events matching
and then we need to trow away the dummies and link with the real events.

So if we find a match, traverse the link of matches and trow away
dummies for that hists.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>