[tools] perf map: Use zalloc for map_groups
Message-id: <20200415143309.16001-190-mpetlan@redhat.com>
Patchwork-id: 302432
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 190/469] perf map: Use zalloc for map_groups
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit ab6cd0e5276e24403751e0b3b8ed807738a8571f
Author: John Keeping <john@metanate.com>
Date: Thu Aug 15 11:01:44 2019 +0100

description
===========
In the next commit we will add new fields to map_groups and we need
these to be null if no value is assigned.  The simplest way to achieve
this is to request zeroed memory from the allocator.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>