[tools] tools lib traceevent: Fix "robust" test of do_generate_dynamic_list_file
Message-id: <20200415143309.16001-198-mpetlan@redhat.com>
Patchwork-id: 302441
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 198/469] tools lib traceevent: Fix "robust" test of do_generate_dynamic_list_file
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit 82a2f88458d70704be843961e10b5cef9a6e95d3
Author: Steven Rostedt (VMware) <rostedt@goodmis.org>
Date: Mon Aug 5 13:01:50 2019 -0400

description
===========
The tools/lib/traceevent/Makefile had a test added to it to detect a failure
of the "nm" when making the dynamic list file (whatever that is). The
problem is that the test sorts the values "U W w" and some versions of sort
will place "w" ahead of "W" (even though it has a higher ASCII value, and
break the test.

Add 'tr "w" "W"' to merge the two and not worry about the ordering.

Fixes: 6467753d61399 ("tools lib traceevent: Robustify do_generate_dynamic_list_file")
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>