[tools] perf top: Fix event group with more than two events
Message-id: <20200415143309.16001-275-mpetlan@redhat.com>
Patchwork-id: 302516
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 275/469] perf top: Fix event group with more than two events
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit be5863b7d9281bbb932542d16b7d758357fde267
Author: Namhyung Kim <namhyung@kernel.org>
Date: Wed Aug 28 08:15:55 2019 +0900

description
===========
The event group feature links relevant hist entries among events so that
they can be displayed together.  During the link process, each hist
entry in non-leader events is connected to a hist entry in the leader
event.  This is done in order of events specified in the command line so
it assumes that events are linked in the order.

But 'perf top' can break the assumption since it does the link process
multiple times.  For example, a hist entry can be in the third event
only at first so it's linked after the leader.  Some time later, second
event has a hist entry for it and it'll be linked after the entry of the
third event.

This makes the code compilicated to deal with such unordered entries.
This patch simply unlink all the entries after it's printed so that they
can assume the correct order after the repeated link process.  Also it'd
be easy to deal with decaying old entries IMHO.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>