[tools] tools lib traceevent: Remove unneeded qsort and uses memmove instead
Message-id: <20200415143309.16001-301-mpetlan@redhat.com>
Patchwork-id: 302544
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 301/469] tools lib traceevent: Remove unneeded qsort and uses memmove instead
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit 301011ba622513cb41ced59973972204e0da2f71
Author: Steven Rostedt (VMware) <rostedt@goodmis.org>
Date: Wed Aug 28 15:05:29 2019 -0400

description
===========
While reading a trace data file that had 100,000s of tasks, the process
took an extremely long time. I profiled it down to add_new_comm(), which
was doing a qsort() call on an array that was pretty much already sorted
(all but the last element. qsort() isn't very efficient when dealing
with mostly sorted arrays, and this definitely showed its issues.

When adding a new task to the task list, instead of using qsort(), do
another bsearch() with a function that will find the element before
where the new task will be inserted in. Then simply shift the rest of
the array, and insert the task where it belongs.

Fixes: f7d82350e597d ("tools/events: Add files to create libtraceevent.a")
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>