[tools] libtraceevent: Round up in tep_print_event() time precision
Message-id: <20200415143309.16001-376-mpetlan@redhat.com>
Patchwork-id: 302619
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 376/469] libtraceevent: Round up in tep_print_event() time precision
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit 5c8da72dc21eff32deb00638e547e2c247ebbfb0
Author: Steven Rostedt (VMware) <rostedt@goodmis.org>
Date: Thu Sep 19 16:51:19 2019 -0400

description
===========
When testing the output of the old trace-cmd compared to the one that
uses the updated tep_print_event() logic, it was different in that the
time stamp precision in the old format would round up to the nearest
precision, where as the new logic truncates. Bring back the old method
of rounding up.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>