[tools] libtraceevent: Man pages fix, rename tep_ref_get() to tep_get_ref()
Message-id: <20200415143309.16001-378-mpetlan@redhat.com>
Patchwork-id: 302621
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 378/469] libtraceevent: Man pages fix, rename tep_ref_get() to tep_get_ref()
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit 38f76c31368b7a8b88659c178f9e620b02660a70
Author: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Date: Thu Sep 19 17:23:37 2019 -0400

description
===========
The tep_ref_get() was renamed to tep_get_ref(), to be more consistent
with the other tep_ref_* APIs. However, in the man pages the API is
still with the old name. The documentation is fixed to reflect the
actual name of the API.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>