[tools] perf docs: Allow man page date to be specified
Message-id: <20200415143309.16001-438-mpetlan@redhat.com>
Patchwork-id: 302681
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 438/469] perf docs: Allow man page date to be specified
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit d586ac10ce56b2381b8e1d8ed74660c1b2b8ab0d
Author: Ian Rogers <irogers@google.com>
Date: Fri Sep 20 21:13:27 2019 -0700

description
===========
With this change if a perf_date parameter is provided to asciidoc then
it will override the default date written to the man page metadata.

Without this change, or if the perf_date isn't specified, then the
current date is written to the metadata.

Having this parameter allows the metadata to be constant if builds
happen on different dates.

The name of the parameter is intended to be consistent with the existing
perf_version parameter.

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>