[tools] perf evlist: Fix fix for freed id arrays
Message-id: <20200415143309.16001-454-mpetlan@redhat.com>
Patchwork-id: 302696
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH 454/469] perf evlist: Fix fix for freed id arrays
Bugzilla: 1818835
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

upstream
========
commit 98a8b2e60c69927b1f405c3b001a1de3f4e53901
Author: Andi Kleen <ak@linux.intel.com>
Date: Fri Oct 11 11:21:40 2019 -0700

description
===========
In the earlier fix for the memory overrun of id arrays I managed to typo
the wrong event in the fix.

Of course we need to close the current event in the loop, not the
original failing event.

The same test case as in the original patch still passes.

Fixes: 7834fa948beb ("perf evlist: Fix access of freed id arrays")
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>