[tools] kvm: selftests: do not blindly clobber registers in guest asm
Message-id: <20200420141722.914884-1-vkuznets@redhat.com>
Patchwork-id: 303628
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 virt PATCH v2 001/614] KVM: selftests: do not blindly clobber registers in guest asm
Bugzilla: 1813987
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Paolo Bonzini <pbonzini@redhat.com>

The guest_code of sync_regs_test is assuming that the compiler will not
touch %r11 outside the asm that increments it, which is a bit brittle.
Instead, we can increment a variable and use a dummy asm to ensure the
increment is not optimized away.  However, we also need to use a
callee-save register or the compiler will insert a save/restore around
the vmexit, breaking the whole idea behind the test.

(Yes, "if it ain't broken...", but I would like the test to be clean
before it is copied into the upcoming s390 selftests).

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 204c91eff798a78498cc7cbf1bc76892badfa96d)
Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>