[kernel] sched/isolation: Allow "isolcpus=" to skip unknown sub-parameters
Message-id: <20200420183353.386622-2-peterx@redhat.com>
Patchwork-id: 304427
Patchwork-instance: patchwork
O-Subject: [RHEL-8.3 PATCH 1/1] sched/isolation: Allow "isolcpus=" to skip unknown sub-parameters
Bugzilla: 1799014
RH-Acked-by: Juri Lelli <juri.lelli@redhat.com>
RH-Acked-by: Ming Lei <ming.lei@redhat.com>
RH-Acked-by: Phil Auld <pauld@redhat.com>

The "isolcpus=" parameter allows sub-parameters before the cpulist is
specified, and if the parser detects an unknown sub-parameters the whole
parameter will be ignored.

This design is incompatible with itself when new sub-parameters are added.
An older kernel will not recognize the new sub-parameter and will
invalidate the whole parameter so the CPU isolation will not take
effect. It emits a warning:

    isolcpus: Error, unknown flag

The better and compatible way is to allow "isolcpus=" to skip unknown
sub-parameters, so that even if new sub-parameters are added an older
kernel will still be able to behave as usual even if with the new
sub-parameter specified on the command line.

Ideally this should have been there when the first sub-parameter for
"isolcpus=" was introduced.

Suggested-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Peter Xu <peterx@redhat.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Link: https://lkml.kernel.org/r/20200403223517.406353-1-peterx@redhat.com

(cherry picked from commit 3662daf023500dc084fa3b96f68a6f46179ddc73)
Signed-off-by: Peter Xu <peterx@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>