[tools] objtool: Move x86 insn decoder to a common location
Message-id: <20200421130652.8998-1-mpetlan@redhat.com>
Patchwork-id: 304460
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH v2 338/469] objtool: Move x86 insn decoder to a common location
Bugzilla: 1818835
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: 1818835
https://bugzilla.redhat.com/show_bug.cgi?id=1818835

v1 --> v2: added "Conflicts:" section

upstream
========
commit d046b725487a97a3a3b35a00e84ca093963b8b4e
Author: Josh Poimboeuf <jpoimboe@redhat.com>
Date: Thu Aug 29 17:41:18 2019 -0500

description
===========
The kernel tree has three identical copies of the x86 instruction
decoder.  Two of them are in the tools subdir.

The tools subdir is supposed to be completely standalone and separate
from the kernel.  So having at least one copy of the kernel decoder in
the tools subdir is unavoidable.  However, we don't need *two* of them.

Move objtool's copy of the decoder to a shared location, so that perf
will also be able to use it.

Conflicts:
==========
Different context in tools/objtool/Makefile due to not (yet) backported
056d28d135bca ("objtool: Query pkg-config for libelf location").

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>