[arm64] crypto: arm64/crct10dif - implement non-Crypto Extensions alternative
Message-id: <20200422233650.3200215-12-msalter@redhat.com>
Patchwork-id: 304583
Patchwork-instance: patchwork
O-Subject: [RHEL 8.3 BZ1826982 PATCH 11/33] crypto: arm64/crct10dif - implement non-Crypto Extensions alternative
Bugzilla: 1826982
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Jeremy Linton <jlinton@redhat.com>

commit 2fffee536c6875bdf546cee0045fed8faa5ea51f
From: Ard Biesheuvel <ard.biesheuvel@linaro.org>

The arm64 implementation of the CRC-T10DIF algorithm uses the 64x64 bit
polynomial multiplication instructions, which are optional in the
architecture, and if these instructions are not available, we fall back
to the C routine which is slow and inefficient.

So let's reuse the 64x64 bit PMULL alternative from the GHASH driver that
uses a sequence of ~40 instructions involving 8x8 bit PMULL and some
shifting and masking. This is a lot slower than the original, but it is
still twice as fast as the current [unoptimized] C code on Cortex-A53,
and it is time invariant and much easier on the D-cache.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Mark Salter <msalter@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>