[arm64] kvm: arm/arm64: Allow user injection of external data aborts
Message-id: <20200503080701.114833-11-gshan@redhat.com>
Patchwork-id: 306046
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 virt PATCH v5 10/75] KVM: arm/arm64: Allow user injection of external data aborts
Bugzilla: 1814009
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Auger Eric <eric.auger@redhat.com>

From: Christoffer Dall <christoffer.dall@arm.com>

In some scenarios, such as buggy guest or incorrect configuration of the
VMM and firmware description data, userspace will detect a memory access
to a portion of the IPA, which is not mapped to any MMIO region.

For this purpose, the appropriate action is to inject an external abort
to the guest.  The kernel already has functionality to inject an
external abort, but we need to wire up a signal from user space that
lets user space tell the kernel to do this.

It turns out, we already have the set event functionality which we can
perfectly reuse for this.

Signed-off-by: Christoffer Dall <christoffer.dall@arm.com>
Signed-off-by: Marc Zyngier <maz@kernel.org>
(cherry picked from commit da345174ceca052469e4775e4ae263b5f27a9355)
Conflicts:
	Documentation/virt/kvm/api.rst
	(changes merged on porting 106ee47dc633)
	include/uapi/linux/kvm.h
	(capability reorder)
Signed-off-by: Gavin Shan <gshan@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>