[arm64] kvm: arm64: Opportunistically turn off WFI trapping when using direct LPI injection
Message-id: <20200503080701.114833-30-gshan@redhat.com>
Patchwork-id: 306065
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 virt PATCH v5 29/75] KVM: arm64: Opportunistically turn off WFI trapping when using direct LPI injection
Bugzilla: 1814009
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Auger Eric <eric.auger@redhat.com>

From: Marc Zyngier <maz@kernel.org>

Just like we do for WFE trapping, it can be useful to turn off
WFI trapping when the physical CPU is not oversubscribed (that
is, the vcpu is the only runnable process on this CPU) *and*
that we're using direct injection of interrupts.

The conditions are reevaluated on each vcpu_load(), ensuring that
we don't switch to this mode on a busy system.

On a GICv4 system, this has the effect of reducing the generation
of doorbell interrupts to zero when the right conditions are
met, which is a huge improvement over the current situation
(where the doorbells are screaming if the CPU ever hits a
blocking WFI).

Signed-off-by: Marc Zyngier <maz@kernel.org>
Reviewed-by: Zenghui Yu <yuzenghui@huawei.com>
Reviewed-by: Christoffer Dall <christoffer.dall@arm.com>
Link: https://lore.kernel.org/r/20191107160412.30301-3-maz@kernel.org
(cherry picked from commit ef2e78ddadbb939ce79553b10dee0131d65d8f3e)
Signed-off-by: Gavin Shan <gshan@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>