[powerpc] kvm: ppc: book3s hv: Don't lose pending doorbell request on migration on P9
Message-id: <20200513005844.107980-8-dgibson@redhat.com>
Patchwork-id: 307284
Patchwork-instance: patchwork
O-Subject: [RHEL-8.3.0 kernel virt PATCH 07/52] KVM: PPC: Book3S HV: Don't lose pending doorbell request on migration on P9
Bugzilla: 1814624
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>

From: Paul Mackerras <paulus@ozlabs.org>

On POWER9, when userspace reads the value of the DPDES register on a
vCPU, it is possible for 0 to be returned although there is a doorbell
interrupt pending for the vCPU.  This can lead to a doorbell interrupt
being lost across migration.  If the guest kernel uses doorbell
interrupts for IPIs, then it could malfunction because of the lost
interrupt.

This happens because a newly-generated doorbell interrupt is signalled
by setting vcpu->arch.doorbell_request to 1; the DPDES value in
vcpu->arch.vcore->dpdes is not updated, because it can only be updated
when holding the vcpu mutex, in order to avoid races.

To fix this, we OR in vcpu->arch.doorbell_request when reading the
DPDES value.

Cc: stable@vger.kernel.org # v4.13+
Fixes: 579006944e0d ("KVM: PPC: Book3S HV: Virtualize doorbell facility on POWER9")
Signed-off-by: Paul Mackerras <paulus@ozlabs.org>
Tested-by: Alexey Kardashevskiy <aik@ozlabs.ru>
(cherry picked from commit ff42df49e75f053a8a6b4c2533100cdcc23afe69)

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1814624

Signed-off-by: David Gibson <dgibson@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>